requirements.txt
streamlit
specklepy
pandas
plotly-express
